vision/HumanFaceKeyPointDetection_FlyAI/requirements.txt
tensorflow
numpy
flyai
Pillow
